Separate cache config tests
