+++ module-ballerina-openapi/Ballerina.toml
@@ -1,10 +1,10 @@
 [package]
 org= "ballerina"
 name= "openapi"
-version= "0.9.0-beta.4"
+version= "@toml.version@"
 
 [[platform.java11.dependency]]
-path = "../openapi-validator/build/libs/openapi-validator-0.9.0-beta.4-SNAPSHOT.jar"
+path = "../openapi-validator/build/libs/openapi-validator-@project.version@.jar"
 groupId = "ballerina"
 artifactId = "openapi"
-version = "0.9.0-beta.4-SNAPSHOT"
+version = "@project.version@"
