+++ openapi-cli/src/test/java/io/ballerina/openapi/generators/openapi/ResponseTests.java
@@ -149,36 +149,6 @@ public class ResponseTests {
         TestUtils.compareWithGeneratedFile(ballerinaFilePath, "response/rs_scenario18.yaml");
     }
 
-    @Test(description = "Generate OpenAPI spec for service configuration annotation in resource without fields")
-    public void cacheConfigTests01() throws OpenApiConverterException, IOException {
-        Path ballerinaFilePath = RES_DIR.resolve("response/configuration_rs.bal");
-        TestUtils.compareWithGeneratedFile(ballerinaFilePath, "response/configuration_rs01.yaml");
-    }
-
-    @Test(description = "When cache-config has custom value without ETag and Last-Modified.")
-    public void cacheConfigTests02() throws OpenApiConverterException, IOException {
-        Path ballerinaFilePath = RES_DIR.resolve("response/configuration_rs02.bal");
-        TestUtils.compareWithGeneratedFile(ballerinaFilePath, "response/cache_config_02.yaml");
-    }
-
-    @Test(description = "When cache-config has custom value with private field and no cache field enable")
-    public void cacheConfigTests03() throws OpenApiConverterException, IOException {
-        Path ballerinaFilePath = RES_DIR.resolve("response/configuration_rs03.bal");
-        TestUtils.compareWithGeneratedFile(ballerinaFilePath, "response/cache_config_03.yaml");
-    }
-
-    @Test(description = "When cache-config has custom value with negative max age")
-    public void cacheConfigTests04() throws OpenApiConverterException, IOException {
-        Path ballerinaFilePath = RES_DIR.resolve("response/configuration_rs04.bal");
-        TestUtils.compareWithGeneratedFile(ballerinaFilePath, "response/cache_config_04.yaml");
-    }
-
-    @Test(description = "When cache-config has custom value with negative max age")
-    public void cacheConfigTests05() throws OpenApiConverterException, IOException {
-        Path ballerinaFilePath = RES_DIR.resolve("response/configuration_rs05.bal");
-        TestUtils.compareWithGeneratedFile(ballerinaFilePath, "response/cache_config_05.yaml");
-    }
-
     @AfterMethod
     public void cleanUp() {
         TestUtils.deleteDirectory(this.tempDir);
