+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/response/cache_config_06.yaml
@@ -1,0 +1,48 @@
+paths:
+  /cachingBackEnd01:
+    get:
+      operationId: operation_get_/cachingBackEnd01
+      responses:
+        "200":
+          description: Ok
+          headers:
+            Cache-Control:
+              schema:
+                type: string
+                default: "must-revalidate,no-store,no-transform,public,proxy-revalidate,max-age=3600"
+            ETag:
+              schema:
+                type: string
+            Last-Modified:
+              schema:
+                type: string
+          content:
+            text/plain:
+              schema:
+                type: string
+        "500":
+          description: Found unexpected output
+  /cachingBackEnd02:
+    get:
+      operationId: operation_get_/cachingBackEnd02
+      responses:
+        "200":
+          description: Ok
+          headers:
+            Cache-Control:
+              schema:
+                type: string
+                default: "must-revalidate,no-store,no-transform,public,proxy-revalidate,max-age=3600,s-maxage=5"
+            ETag:
+              schema:
+                type: string
+            Last-Modified:
+              schema:
+                type: string
+          content:
+            text/plain:
+              schema:
+                type: string
+        "500":
+          description: Found unexpected output
+components: {}
