+++ openapi-cli/src/test/resources/testng.xml
@@ -34,6 +34,7 @@ under the License.
             <class name="io.ballerina.openapi.generators.openapi.ListenerTests"/>
             <class name="io.ballerina.openapi.generators.openapi.ModuleReferenceTests"/>
             <class name="io.ballerina.openapi.generators.openapi.HeaderTests"/>
+            <class name="io.ballerina.openapi.generators.openapi.CacheConfigTests"/>
             <class name="io.ballerina.openapi.cmd.OpenApiGenServiceCmdTest"/>
             <class name="io.ballerina.openapi.CodeGeneratorTest"/>
             <class name="io.ballerina.openapi.generators.common.GeneratorUtilsTests"/>
